totally rebuilded whole backend removed half of the stuff now this gonna base on typescript
