+++ .gitignore
@@ -1,3 +1,8 @@
 node_modules
 .env
+config.ts
 config
+@types
+module.d.ts
+dist
+.vscode
