+++ package.json
@@ -2,31 +2,41 @@
   "name": "backend",
   "version": "1.0.0",
   "description": "",
-  "main": "server.js",
+  "main": "server.ts",
   "type": "module",
   "scripts": {
+    "start": "tsx --watch src/server.ts",
     "test": "vitest",
-    "start": "node server.js",
-    "dev": " nodemon --ignore 'logs/*' server.js"
+    "build": "tsc"
   },
   "keywords": [],
   "author": "",
   "license": "ISC",
   "dependencies": {
+    "@types/dotenv": "^8.2.0",
     "bcrypt": "^5.1.1",
     "cors": "^2.8.5",
-    "dotenv": "^16.3.1",
+    "dotenv": "^16.4.5",
     "express": "^4.18.2",
-    "express-session": "^1.18.0",
     "Joi": "^17.11.0",
     "jsonwebtoken": "^9.0.2",
     "mongoose": "^7.4.4",
     "nodemailer": "^6.9.11",
-    "nodemon": "^3.0.1"
+    "swagger-jsdoc": "^6.2.8",
+    "swagger-ui-express": "^5.0.0"
   },
   "devDependencies": {
     "@babel/core": "^7.23.3",
     "@babel/preset-env": "^7.23.3",
+    "@types/bcrypt": "^5.0.2",
+    "@types/cors": "^2.8.17",
+    "@types/express": "^4.17.21",
+    "@types/express-session": "^1.18.0",
+    "@types/jsonwebtoken": "^9.0.6",
+    "@types/node": "^20.12.10",
+    "@types/nodemailer": "^6.4.15",
+    "tsx": "^4.9.3",
+    "typescript": "^5.4.5",
     "vitest": "^0.34.6"
   }
 }
