--- src/auth/OTP/otp.constants.js
@@ -1,4 +1,0 @@
-const OTP_EXPIRY_MINUTES = 15;
-const expiresAt = new Date(Date.now() + OTP_EXPIRY_MINUTES * 60 * 1000);
-
-export { expiresAt };
