--- src/auth/OTP/otp.model.js
@@ -1,19 +1,0 @@
-import mongoose, { Schema } from "mongoose";
-
-const otpSchema = new Schema({
-  email: {
-    type: String,
-    required: true,
-  },
-  otp: {
-    type: String,
-    required: true,
-  },
-  expiresAt: {
-    type: Date,
-    required: true,
-  },
-});
-
-const OTP = mongoose.model("OTP", otpSchema);
-export { OTP };
