+++ src/email/mail.controller.ts
@@ -1,7 +1,11 @@
 import nodemailer from "nodemailer";
-import { EMAIL_PASS, EMAIL_USER } from "../../config/config.js";
-
-export const sendMail = async (to, subject, text, html) => {
+import { EMAIL_PASS, EMAIL_USER } from "../config/config.js";
+export const sendMail = async (
+  to: string,
+  subject: string,
+  text: string,
+  html: string
+) => {
   const transporter = nodemailer.createTransport({
     host: "smtp.gmail.com",
     port: 587,
